+++ ringbuffer.h
@@ -8,16 +8,16 @@ class RingBuffer : public QObject
     Q_OBJECT
 public:
     //explicit RingBuffer(QObject *parent = 0);
-   explicit  RingBuffer( unsigned int buffSize,QObject *parent = 0 );
+   explicit  RingBuffer( qint64 buffSize,QObject *parent = 0 );
     qint64 addData(const void* data,  qint64 len );
     qint64 getData(void* outData, qint64 len );
-    void skipData( unsigned int& len );
+    void skipData( qint64& len );
     // The amount of data the buffer can currently receive on one addData() call.
-    inline unsigned int Free_Space()    { return (unsigned int)_bufferSize; }
+    inline qint64 Free_Space()    { return (qint64)_bufferSize; }
 
     // The total amount of data in the buffer. Note that it may not be continuous: you may need
     // two successive calls to getData() to get it all.
-    inline unsigned int Buffered_Bytes() { return (unsigned int)_dataInBuffer; }
+    inline qint64 Buffered_Bytes() { return (qint64)_dataInBuffer; }
 signals:
 
 public slots:
@@ -28,7 +28,7 @@ private:
     char *_consumePointer;
     char *_buffEnd;
     char *_buffer;
-    int _bufferSize, _maxConsume, _dataInBuffer;
+    qint64 _bufferSize, _maxConsume, _dataInBuffer;
